refactor(e2e): unify RouterInstance into Gateway class, split conftest.py into modular fixtures
Add cloud mode support to Gateway class and remove the duplicate
RouterInstance dataclass from backends.py. The Gateway class now
supports four modes:
1. Regular mode: worker URLs + model path
2. PD mode: prefill/decode workers
3. IGW mode: empty start, add workers via API
4. Cloud mode: OpenAI/xAI runtimes

Also clarify the distinction between:
- Cloud runtimes: Where models run (openai, xai)
- History backends: Gateway plugin for storage (memory, oracle)

Backward compatibility maintained for e2e_response_api migration.

Split the 1056-line conftest.py into focused modules:
- conftest.py (192 lines): Path setup, logging, imports
- fixtures/hooks.py (358 lines): Pytest collection hooks, GPU validation
- fixtures/pool.py (194 lines): Model pool fixtures
- fixtures/setup_backend.py (401 lines): Backend setup fixtures
- fixtures/markers.py (57 lines): Marker helper utilities

Also marked legacy fixtures (ports.py, router_manager.py) as deprecated
for removal during e2e_response_api migration.
